video_generator_app/requirements.txt
google-cloud-aiplatform>=1.25.0
google-cloud-texttospeech>=2.14.1
moviepy>=1.0.3
Pillow>=9.0.0
customtkinter>=5.1.2
python-dotenv>=0.21.0
